🔨 resrtucture repo, move abstract class to abstract/ dir
