+++ abstracts/base_entry_strategy.py
