+++ abstracts/base_exit_strategy.py
