+++ abstracts/base_live_trade_client.py
@@ -1,4 +1,4 @@
-from trade_clients.base_tarde_client import BaseTradeClient
+from abstracts.base_tarde_client import BaseTradeClient
 
 class BaseLiveTradeClient(BaseTradeClient):
     def __init__(self) -> None:
