+++ trade_clients/offline/offline_live_client.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from trade_clients.base_live_trade_client import BaseLiveTradeClient
+from abstracts.base_live_trade_client import BaseLiveTradeClient
 
 MOCK_KLINES_PATH = 'resources/mock_klines.json'
 
@@ -11,3 +11,4 @@ class OfflineLiveTradeClient(BaseLiveTradeClient):
     def fetch_klines(self, symbol, timeframe, timeframe_limit=100):
         return pd.read_json(MOCK_KLINES_PATH, orient='records', lines=True)
         
+# EOF
